// bdt/src/components/baseData/meter/DeviceCategory/treeview/treeDropdown/interface.ts
import { TypeDoMainStore } from "../domainStore";

/**
 * 设备种类/类型树的Props
 */
export interface ITreedownselectProps {
    /**
     * 数据源
     */
    list?: any;

    /**
     * 改变事件
     */
    onChange?: (value: any, label: any) => void;

    /**
     * 当前选择项
     */
    value?: any;

    /***
     * 选中事件
     */
    onSelect?: (fatherId: string) => void;
    /**
     * 设置字段值，表单使用
     */
    setFieldsValue?: (obj: object) => void;

    /**
     * Store
     */
    TreeTypeStore?: TypeDoMainStore
}